fix(api): bad implementation

+++ app/src/main/java/com/joshuatabakhoff/caddik/network/model/Ingredient.kt
@@ -7,8 +7,8 @@ class Ingredient {
     var text = ""
 
     var has_sub_ingredients = ""
-    var percent_min: Int? = 0
-    var percent_max: Int? = 100
+    var percent_min: Double? = 0.0
+    var percent_max: Double? = 100.0
 
     var vegetarian: String? = "yes"
     var vegan: String? = "yes"
